Fetch new forum post on clicking reload button
